tc_build: Introduce a common GitSourceManager
We will be adding Rust support in the future, so pull out common git
repository management code from LLVMSourceManager that can be shared
with a RustSourceManager.
